@@ -32,11 +32,12 @@
     "popper.js": "^1.16.1",
     "postcss": "^8.1.0",
     "propagating-hammerjs": "^3.0.0",
-    "react": "^16.13.1",
-    "react-app-polyfill": "^1.0.6",
-    "react-dom": "^16.13.1",
-    "react-grid-layout": "^0.18.3",
-    "react-modal": "^3.11.2",
+    "react": "^19.1.1",
+    "react-app-polyfill": "^3.0.0",
+    "react-bootstrap": "^2.10.10",
+    "react-dom": "^19.1.1",
+    "react-grid-layout": "^1.5.2",
+    "react-modal": "^3.16.3",
     "regenerator-runtime": "^0.14.1",
     "summernote": "^0.9.1",
     "tippy.js": "^6.3.7",
@@ -57,14 +58,16 @@
     "@jest/globals": "^30.0.5",
     "@stylistic/eslint-plugin": "^5.2.0",
     "@testing-library/dom": "^10.0.0",
-    "@testing-library/react": "12",
+    "@testing-library/react": "16.3.0",
+    "@types/form-serialize": "^0.7.4",
     "@types/jest": "^30.0.0",
     "@types/jquery": "^3.5.32",
     "@types/jstree": "^3.3.46",
     "@types/node": "^24.2.0",
-    "@types/react": "^17.0.41",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^19.1.9",
+    "@types/react-dom": "^19.1.7",
     "@types/react-grid-layout": "^1.3.2",
+    "@types/react-modal": "^3.16.3",
     "@types/typeahead.js": "^0.11.6",
     "autoprefixer": "^10.4.21",
     "babel-loader": "^10.0.0",
