@@ -1,23 +1,23 @@
-import { initializeComponent, getComponentElements } from 'component';
+import { initializeComponent, getComponentElements } from 'component'
 
 export default (scope) => {
 
-    if (!getComponentElements(scope, '.dashboard').length) {
-        return;
-    }
+  if (!getComponentElements(scope, '.dashboard').length) {
+    return;
+  }
 
-    import(
+  import(
     /* webpackChunkName: "dashboard" */
-        './lib/component'
+    './lib/component'
+  ).then(({ default: Component }) => {
+    initializeComponent(scope, '.dashboard', Component)
+  }).then(() => {
+    import(
+      /* webpackChunkName: "dashboardgraph" */
+      './dashboard-graph/lib/component'
     ).then(({ default: Component }) => {
-        initializeComponent(scope, '.dashboard', Component);
-    }).then(() => {
-        import(
-            /* webpackChunkName: "dashboardgraph" */
-            './dashboard-graph/lib/component'
-        ).then(({ default: Component }) => {
-            initializeComponent(scope, '.dashboard-graph', Component);
-        });
-    });
+      initializeComponent(scope, '.dashboard-graph', Component)
+    })
+  });
 
-};
+}
