@@ -4,12 +4,13 @@ import '@testing-library/dom';
 import { describe, it, expect, jest } from '@jest/globals';
 
 import Footer from './Footer';
+import { FooterProps } from './types';
 
 import 'testing/extensions';
 
 describe('Footer', () => {
     it('Creates a footer', () => {
-        const footerProps = {
+        const footerProps: FooterProps = {
             addWidget: jest.fn(),
             currentDashboard: {
                 name: 'Dashboard 1',
@@ -24,16 +25,16 @@ describe('Footer', () => {
         render(<Footer {...footerProps} />);
 
         expect(screen.getByText('Download')).toBeInstanceOf(HTMLButtonElement);
-        act(() => screen.getByText('Download').click());
-        // @ts-expect-error Extension function
+        act(()=>screen.getByText('Download').click());
+        // @ts-expect-error extension method
         expect(screen.getByText('As PDF')).toHaveAttribute('href', 'http://localhost:3000/dashboard/1/download');
-        act(() => screen.getByText('Add Widget').click());
+        act(()=>screen.getByText('Add Widget').click());
         expect(screen.getByText('type1')).toBeInstanceOf(HTMLAnchorElement);
         expect(screen.getByText('type2')).toBeInstanceOf(HTMLAnchorElement);
     });
 
     it('Creates a footer without download', () => {
-        const footerProps = {
+        const footerProps: FooterProps = {
             addWidget: jest.fn(),
             currentDashboard: {
                 name: 'Dashboard 1',
@@ -51,7 +52,7 @@ describe('Footer', () => {
     });
 
     it('Creates a footer with read only', () => {
-        const footerProps = {
+        const footerProps: FooterProps = {
             addWidget: jest.fn(),
             currentDashboard: {
                 name: 'Dashboard 1',
@@ -69,7 +70,7 @@ describe('Footer', () => {
     });
 
     it('Creates a footer with no download and read only', () => {
-        const footerProps = {
+        const footerProps: FooterProps = {
             addWidget: jest.fn(),
             currentDashboard: {
                 name: 'Dashboard 1',
@@ -87,9 +88,9 @@ describe('Footer', () => {
         expect(screen.queryByText('Add Widget')).toBeNull();
     });
 
-    it('Clicks the create modal button with expected parameters', () => {
+    it('Clicks the create modal button with expected parameters', ()=>{
         const addWidget = jest.fn();
-        const footerProps = {
+        const footerProps: FooterProps = {
             addWidget,
             currentDashboard: {
                 name: 'Dashboard 1',
@@ -103,16 +104,16 @@ describe('Footer', () => {
 
         render(<Footer {...footerProps} />);
 
-        act(() => screen.getByText('Add Widget').click());
+        act(()=>screen.getByText('Add Widget').click());
         const type1WidgetButton = screen.getByText('type1');
-        act(() => type1WidgetButton.click());
+        act(()=>type1WidgetButton.click());
 
         expect(addWidget).toHaveBeenCalledTimes(1);
         expect(addWidget).toHaveBeenCalledWith('type1');
         addWidget.mockClear();
 
         const addWidgetButton2 = screen.getByText('type2');
-        act(() => addWidgetButton2.click());
+        act(()=>addWidgetButton2.click());
 
         expect(addWidget).toHaveBeenCalledTimes(1);
         expect(addWidget).toHaveBeenCalledWith('type2');
